Fix in versioned API if the version contains a dot
In version 3.3.0 of the task, if we create a new version for a given
API we concat the name of the api with the version. For example:

```powershell
$name="my-api"
$version="1.0"
$url="${name}${version}"
```

Microsoft only allows letters, numbers and hyphens in their identifiers.
Therefore, we've to get rid of the `.` in the id.

As a first solution, we now replace `.` with `-` in the identifier of
the API version.

Fixes #49